+++ class/index.js
@@ -14,6 +14,7 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
     for (var p in m) if (p !== "default" && !Object.prototype.hasOwnProperty.call(exports, p)) __createBinding(exports, m, p);
 };
 exports.__esModule = true;
+__exportStar(require("./all-or-none.type"), exports);
 __exportStar(require("./complete.type"), exports);
 __exportStar(require("./concrete.type"), exports);
 __exportStar(require("./constructor.type"), exports);