+++ class/index.ts
@@ -1,3 +1,4 @@
+export * from "./all-or-none.type";
 export * from "./complete.type";
 export * from "./concrete.type";
 export * from "./constructor.type";