+++ string/uuid.type.d.ts
@@ -16,5 +16,6 @@ export declare type UUIDType = string & {
 export declare type UUID = string & {
     readonly __uuid__: void;
 };
-export declare function isUuid(inputUUID: string): inputUUID is TypeUUID;
-export declare function assertUuid(inputUUID: unknown): TypeUUID;
+export declare type UUIDVersionType = '3' | '4' | '5' | 'all' | 3 | 4 | 5;
+export declare function isUuid(inputUUID: unknown, version?: UUIDVersionType): inputUUID is TypeUUID;
+export declare function assertUuid(inputUUID: unknown, version?: UUIDVersionType): TypeUUID;